requirements.txt
torch
torchvision
flask
flask-cors
Pillow
pandas
kagglehub